requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.11.28
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
et-xmlfile==1.0.1
fa2==0.3.5
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
joblib==0.13.2
kiwisolver==1.1.0
leidenalg==0.7.0
llvmlite==0.29.0
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
mysqlclient==1.3.14
neo4j-driver==1.7.4
neobolt==1.7.13
neotime==1.7.4
networkx==2.3
numba==0.45.1
numpy==1.16.4
openpyxl==2.6.3
ortools==7.1.6720
pandas==0.25.1
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pycairo==1.18.1
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
python-igraph==0.7.1.post7
python-louvain==0.13
pytz==2019.2
PyYAML==5.1.2
scikit-learn==0.21.2
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
statsmodels==0.10.1
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
wcwidth==0.1.7
